added scheduling logic, refactoring

+++ src/main/java/com/example/Bank/model/Credit.java
@@ -6,8 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @AllArgsConstructor
 @NoArgsConstructor
 @Setter
@@ -18,6 +16,6 @@ public class Credit {
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id", nullable = false)
     private Long id;
-    private BigDecimal creditLimit;
+    private double creditLimit;
     private double interestRate;
 }
